Anti-fancification on sampling
'proper' sampling of background to do errnat reading reconstruction was proving to be impractical, and lack of inter-operability with laserpp limited the time available, so it was taken out in foavor of a short-term brute force approach: replace errant (0.0) readings with a maximul range for the product.   This will require laserpp to have inclusion threshold set beyond this 20m range.